Imporve the param type in li func
The li function just accept the int32_t type param, but the li in lock_object give an
unsigned long type, so just improve it.
The bishengjdk use the int64_t in this place. This patch will use the int32_t type.

+++ src/hotspot/cpu/riscv32/interp_masm_riscv32.cpp
@@ -819,7 +819,7 @@ void InterpreterMacroAssembler::lock_object(Register lock_reg)
     // least significant 3 bits clear.
     // NOTE: the oopMark is in swap_reg %x10 as the result of cmpxchg
     sub(swap_reg, swap_reg, sp);
-    li(t0, (unsigned long)(7 - os::vm_page_size()));
+    li(t0, (int32_t)(7 - os::vm_page_size()));
     andr(swap_reg, swap_reg, t0);
 
     // Save the test result, for recursive case, the result is zero
